Move the fizz action processing to FizzServerHandshake
