+++ quic/server/handshake/ServerHandshake.cpp
@@ -239,172 +239,13 @@ void ServerHandshake::processPendingEvents() {
   }
 }
 
-class ServerHandshake::ActionMoveVisitor : public boost::static_visitor<> {
- public:
-  explicit ActionMoveVisitor(ServerHandshake& server) : server_(server) {}
-
-  void operator()(fizz::DeliverAppData&) {
-    server_.onError(std::make_pair(
-        "Unexpected data on crypto stream",
-        TransportErrorCode::PROTOCOL_VIOLATION));
-  }
-
-  void operator()(fizz::WriteToSocket& write) {
-    server_.onWriteData(write);
-  }
-
-  void operator()(fizz::server::ReportEarlyHandshakeSuccess&) {
-    server_.phase_ = Phase::KeysDerived;
-  }
-
-  void operator()(fizz::server::ReportHandshakeSuccess&) {
-    server_.handshakeDone_ = true;
-    auto originalPhase = server_.phase_;
-    // Fizz only reports handshake success when the server receives the full
-    // client finished. At this point we can write any post handshake data and
-    // crypto data with the 1-rtt keys.
-    server_.phase_ = Phase::Established;
-    if (originalPhase != Phase::Handshake) {
-      // We already derived the zero rtt keys as well as the one rtt write
-      // keys.
-      server_.onHandshakeDone();
-    }
-  }
-
-  void operator()(fizz::ReportError& err) {
-    auto errMsg = err.error.what();
-    if (errMsg.empty()) {
-      errMsg = "Error during handshake";
-    }
-
-    auto fe = err.error.get_exception<fizz::FizzException>();
-
-    if (fe && fe->getAlert()) {
-      auto alertNum =
-          static_cast<std::underlying_type<TransportErrorCode>::type>(
-              fe->getAlert().value());
-      alertNum += static_cast<std::underlying_type<TransportErrorCode>::type>(
-          TransportErrorCode::CRYPTO_ERROR);
-      server_.onError(std::make_pair(
-          errMsg.toStdString(), static_cast<TransportErrorCode>(alertNum)));
-    } else {
-      server_.onError(std::make_pair(
-          errMsg.toStdString(),
-          static_cast<TransportErrorCode>(
-              fizz::AlertDescription::internal_error)));
-    }
-  }
-
-  void operator()(fizz::WaitForData&) {
-    server_.waitForData_ = true;
-  }
-
-  void operator()(fizz::server::MutateState& mutator) {
-    mutator(server_.state_);
-  }
-
-  void operator()(fizz::server::AttemptVersionFallback&) {
-    CHECK(false) << "Fallback Unexpected";
-  }
-
-  void operator()(fizz::EndOfData&) {
-    server_.onError(std::make_pair(
-        "Unexpected close notify received",
-        TransportErrorCode::INTERNAL_ERROR));
-  }
-
-  void operator()(fizz::SecretAvailable& secretAvailable) {
-    switch (secretAvailable.secret.type.type()) {
-      case fizz::SecretType::Type::EarlySecrets_E:
-        switch (*secretAvailable.secret.type.asEarlySecrets()) {
-          case fizz::EarlySecrets::ClientEarlyTraffic:
-            server_.computeCiphers(
-                CipherKind::ZeroRttRead,
-                folly::range(secretAvailable.secret.secret));
-            break;
-          default:
-            break;
-        }
-        break;
-      case fizz::SecretType::Type::HandshakeSecrets_E:
-        switch (*secretAvailable.secret.type.asHandshakeSecrets()) {
-          case fizz::HandshakeSecrets::ClientHandshakeTraffic:
-            server_.computeCiphers(
-                CipherKind::HandshakeRead,
-                folly::range(secretAvailable.secret.secret));
-            break;
-          case fizz::HandshakeSecrets::ServerHandshakeTraffic:
-            server_.computeCiphers(
-                CipherKind::HandshakeWrite,
-                folly::range(secretAvailable.secret.secret));
-            break;
-        }
-        break;
-      case fizz::SecretType::Type::AppTrafficSecrets_E:
-        switch (*secretAvailable.secret.type.asAppTrafficSecrets()) {
-          case fizz::AppTrafficSecrets::ClientAppTraffic:
-            server_.computeCiphers(
-                CipherKind::OneRttRead,
-                folly::range(secretAvailable.secret.secret));
-            break;
-          case fizz::AppTrafficSecrets::ServerAppTraffic:
-            server_.computeCiphers(
-                CipherKind::OneRttWrite,
-                folly::range(secretAvailable.secret.secret));
-            break;
-        }
-        break;
-      case fizz::SecretType::Type::MasterSecrets_E:
-        break;
-    }
-    server_.handshakeEventAvailable_ = true;
-  }
-
- private:
-  ServerHandshake& server_;
-};
-
 void ServerHandshake::processActions(
     fizz::server::ServerStateMachine::CompletedActions actions) {
   // This extra DestructorGuard is needed due to the gap between clearing
   // actionGuard_ and potentially processing another action.
   folly::DelayedDestruction::DestructorGuard dg(conn_);
 
-  ActionMoveVisitor visitor(*this);
-  for (auto& action : actions) {
-    switch (action.type()) {
-      case fizz::server::Action::Type::DeliverAppData_E:
-        visitor(*action.asDeliverAppData());
-        break;
-      case fizz::server::Action::Type::WriteToSocket_E:
-        visitor(*action.asWriteToSocket());
-        break;
-      case fizz::server::Action::Type::ReportHandshakeSuccess_E:
-        visitor(*action.asReportHandshakeSuccess());
-        break;
-      case fizz::server::Action::Type::ReportEarlyHandshakeSuccess_E:
-        visitor(*action.asReportEarlyHandshakeSuccess());
-        break;
-      case fizz::server::Action::Type::ReportError_E:
-        visitor(*action.asReportError());
-        break;
-      case fizz::server::Action::Type::EndOfData_E:
-        visitor(*action.asEndOfData());
-        break;
-      case fizz::server::Action::Type::MutateState_E:
-        visitor(*action.asMutateState());
-        break;
-      case fizz::server::Action::Type::WaitForData_E:
-        visitor(*action.asWaitForData());
-        break;
-      case fizz::server::Action::Type::AttemptVersionFallback_E:
-        visitor(*action.asAttemptVersionFallback());
-        break;
-      case fizz::server::Action::Type::SecretAvailable_E:
-        visitor(*action.asSecretAvailable());
-        break;
-    }
-  }
+  processCryptoActions(std::move(actions));
 
   actionGuard_ = folly::DelayedDestruction::DestructorGuard(nullptr);
   if (callback_ && !inHandshakeStack_ && handshakeEventAvailable_) {
